+++ src/routes/main.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState, ChangeEvent, FormEvent } from 'react';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
+import { axiosInstance, useTokenRefresher } from '../utils/apiInstance';
 
 import '../index.css';
 
@@ -19,54 +20,57 @@ interface ApiResponse {
 
 function Main() {
   const planStatus: string = '플랜 이용중';
-  const accessToken = localStorage.getItem('accessToken');
+  // const accessToken = localStorage.getItem('accessToken');
   // const refreshToken = localStorage.getItem('refreshToken');
   const [data, setData] = useState<ApiResponse | null>(null);
   const [searchInputValue, setSearchInputValue] = useState('');
 
-  useEffect(() => {
-    axios
-      .get(`${import.meta.env.VITE_API_URL}/me/summary`, {
-        headers: {
-          accept: 'application/json',
-          Authorization: `Bearer ${accessToken}`,
-        },
-      })
-      .then((response) => {
-        // console.log(response.data);
-        setData(response.data);
-      });
-  }, [accessToken]);
+  // useEffect(() => {
+  //   axios
+  //     .get(`${import.meta.env.VITE_API_URL}/me/summary`, {
+  //       headers: {
+  //         accept: 'application/json',
+  //         Authorization: `Bearer ${accessToken}`,
+  //       },
+  //     })
+  //     .then((response) => {
+  //       // console.log(response.data);
+  //       setData(response.data);
+  //     });
+  // }, [accessToken]);
 
-  useEffect(() => {
-    const interval = setInterval(
-      () => {
-        axios
-          .post(
-            `${import.meta.env.VITE_API_URL}/tokens`,
-            {},
-            {
-              headers: {
-                accept: 'application/json',
-                Authorization: `Bearer ${localStorage.getItem('refreshToken')}`,
-              },
-            },
-          )
-          .then((response) => {
-            localStorage.setItem('accessToken', response.data.accessToken);
-            localStorage.setItem('refreshToken', response.data.refreshToken);
-          })
-          .catch((error) => {
-            console.error(error);
-          });
-      },
-      15 * 60 * 1000,
-    );
+  axiosInstance.get('/me/summary');
 
-    return () => {
-      clearInterval(interval);
-    };
-  }, []);
+  // useEffect(() => {
+  //   const interval = setInterval(
+  //     () => {
+  //       axios
+  //         .post(
+  //           `${import.meta.env.VITE_API_URL}/tokens`,
+  //           {},
+  //           {
+  //             headers: {
+  //               accept: 'application/json',
+  //               Authorization: `Bearer ${localStorage.getItem('refreshToken')}`,
+  //             },
+  //           },
+  //         )
+  //         .then((response) => {
+  //           localStorage.setItem('accessToken', response.data.accessToken);
+  //           localStorage.setItem('refreshToken', response.data.refreshToken);
+  //         })
+  //         .catch((error) => {
+  //           console.error(error);
+  //         });
+  //     },
+  //     15 * 60 * 1000,
+  //   );
+
+  //   return () => {
+  //     clearInterval(interval);
+  //   };
+  // }, []);
+  useTokenRefresher();
 
   const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault();
